% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasteR.R
\name{fasteR}
\alias{fasteR}
\title{TopicFundeR}
\usage{
fasteR(x, reqtime = 0, id_only = FALSE)
}
\arguments{
\item{x}{A name put in quotation marks.}

\item{reqtime}{A integer number specifying the seconds to wait between requests.
Default is set to 0.}

\item{id_only}{If TRUE, returns only a vector containg the ID.}
}
\value{
Returns a dataframe consisting of a ID-Variable, the name-Variable
    and the input-Variable.
}
\description{
This function has the same properties as "findeR" except that
    it takes an additional parameter (id_only)
}
\examples{
names <- c("name1", "name2", "name3", "...")
id <- sapply(names, fasteR, reqtime = 5, id_only = TRUE)


}
