% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_info.R
\name{find_info}
\alias{find_info}
\title{PackageMA - find_info}
\usage{
find_info(x, reqtime = 0, index = TRUE)
}
\arguments{
\item{x}{A numeric ID.}

\item{reqtime}{An integer number specifying the number of seconds to wait between requests.
Default is set to 0.}

\item{index}{If set to TRUE, a continuous index keeps track of the ID's that were already scraped.}
}
\value{
Returns a dataframe with additional project variables, such as the number of
     projects, the name of the projects and their corresponding IDs.
}
\description{
Extract additional information based on a person's ID.
}
\examples{
 find_info(id)
x <- fasteR("Jürgen Gerhards")
y <- find_info(x$id)
z <- dplyr::left_join(y, x)


}
