% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_info.R
\name{find_info2}
\alias{find_info2}
\title{TopicFundeR}
\usage{
find_info2(x, reqtime = 0)
}
\arguments{
\item{x}{A name put in quotation marks.}

\item{reqtime}{A integer number specifying the seconds to wait between requests.
Default is set to 0.}
}
\value{
Returns a dataframe consisting of a ID-Variable, the name-Variable
    and the input-Variable
}
\description{
Search the GEPRIS-Database for a name and extract the information.
   FindeR and find_info are best used for single names and IDs. For multiple
   values use fasteR and wrap_it for convenience.
}
\examples{
findeR("Name Surname", reqtime = 5)
findeR("Name")

}
