% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_fun.R
\name{meta_fun}
\alias{meta_fun}
\title{PackageMA - meta_fun}
\usage{
meta_fun(x, reqtime = 0)
}
\arguments{
\item{x}{A character name.}

\item{reqtime}{An integer number specifying the number of seconds to wait between requests.
Default is set to 0.}
}
\value{
Returns a dataframe with all available information.
}
\description{
Search the GEPRIS-Database for a name and extract the information.
    meta_fun is performing all the other functions at once without interim steps.
}
\examples{

df <- meta_fun("J*rgen Gerhards", reqtime = 5)


}
