% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pasteR.R
\name{pasteR}
\alias{pasteR}
\title{TopicFundeR}
\usage{
pasteR(x, lengthline = 65)
}
\arguments{
\item{x}{A string or a input coercible to one}

\item{lengthline}{integer describing the wanted number of characters in each line}
}
\description{
Paste long strings neatly to console in order to fit the limits
    of your screen
}
\examples{
pasteR(long_link)

}
