% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotteR.R
\name{plotteR}
\alias{plotteR}
\title{TopicFundeR}
\usage{
plotteR(df, bipartite = FALSE, plotting = TRUE)
}
\arguments{
\item{bipartite}{If TRUE returns a two-mode network, otherwise a one-mode network}

\item{plotting}{If TRUE the network will be plotted (default option). Otherwise the two-mode-graph is returned.}

\item{x}{A Dataframe.}
}
\value{
Plots a Network to viewer.
}
\description{
Plots a rudimental network. Either a bipartite or normal one.
}
\examples{
plotteR(df)

}
