% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps.R
\name{steps}
\alias{steps}
\title{TopicFundeR}
\usage{
steps(project_id, reqtime = 0)
}
\arguments{
\item{reqtime}{A integer number specifying the seconds to wait between requests.
Default is set to 0.}

\item{x}{A numeric ID.}
}
\value{
Returns a dataframe consisting of informations about all joined in scientists.
}
\description{
Takes a project ID and extracts the involved scientists. Wraps
    around wrap_it.
}
\examples{
 df <- steps(id, reqtime = 2)


}
