% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/steps2.R
\name{steps2}
\alias{steps2}
\title{TopicFundeR}
\usage{
steps2(project_id, reqtime = 0, texts = FALSE, projects = FALSE)
}
\arguments{
\item{reqtime}{A integer number specifying the seconds to wait between requests.
Default is set to 0.}

\item{x}{A numeric ID.}
}
\value{
Returns a dataframe consisting of informations about all joined in scientists and conditionally project-texts.
}
\description{
Takes a project ID or ID of a person and extracts the involved scientists. Wraps
    around wrap_it.
}
\examples{
 df <- steps2(person_id, reqtime = 2, texts = TRUE, projects = FALSE)
 df <- steps2(project_id, reqtime = 2, texts = TRUE, projects = TRUE)

}
