% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_info2.R
\name{find_info2}
\alias{find_info2}
\title{TopicFundeR}
\usage{
find_info2(x, reqtime = 0)
}
\arguments{
\item{x}{A numeric ID.}

\item{reqtime}{A integer number specifying the seconds to wait between requests.
Default is set to 0.}
}
\value{
Returns a dataframe with additional project variables, such as number of
     projects, name of the projects and their corresponding IDs.
}
\description{
Extract additional information based on a person's ID.
     FindeR and find_info are best used for single names and IDs. For multiple
     values use fasteR and wrap_it for convenience.
}
\examples{
 find_info2(id)
x <- findeR("name")
y <- find_info(x$id)
z <- dplyr::left_join(y, x)


}
