% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotteR.R
\name{plotteR}
\alias{plotteR}
\title{TopicFundeR}
\usage{
plotteR(df, bipartite = FALSE)
}
\arguments{
\item{x}{A Dataframe.}
}
\value{
Plots a Network to viewer.
}
\description{
Plots a rudimental network. Either a bipartite or normal one.
}
\examples{
plotteR(df)

}
