% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_it.R
\name{wrap_it}
\alias{wrap_it}
\title{TopicFundeR}
\usage{
wrap_it(ids, reqtime = 0, index = TRUE)
}
\arguments{
\item{ids}{Vector containing IDs}

\item{reqtime}{A integer number specifying the seconds to wait between requests.
Default is set to 0. Recommended is 5.}

\item{index}{If set to TRUE, a continuous index keeps track of IDs already scraped.}
}
\value{
Returns a dataframe consisting of a ID-Variable, the name-Variable
    and the input-Variable and several others.
}
\description{
Used for extracting information of multiple IDs.
}
\examples{
wrap_it(ids, reqtime = 5, index = TRUE)

}
